# apps/api/internal/api/types/types.gen.go
// Package types provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package types

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for CrawlStatus.
const (
	Queued CrawlStatus = "queued"
)

// Valid indicates whether the value is a known member of the CrawlStatus enum.
func (e CrawlStatus) Valid() bool {
	switch e {
	case Queued:
		return true
	default:
		return false
	}
}

// Defines values for HealthStatusStatus.
const (
	HealthStatusStatusOk HealthStatusStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatusStatus enum.
func (e HealthStatusStatus) Valid() bool {
	switch e {
	case HealthStatusStatusOk:
		return true
	default:
		return false
	}
}

// Defines values for ReadinessStatusChecksDb.
const (
	ReadinessStatusChecksDbDown ReadinessStatusChecksDb = "down"
	ReadinessStatusChecksDbOk   ReadinessStatusChecksDb = "ok"
)

// Valid indicates whether the value is a known member of the ReadinessStatusChecksDb enum.
func (e ReadinessStatusChecksDb) Valid() bool {
	switch e {
	case ReadinessStatusChecksDbDown:
		return true
	case ReadinessStatusChecksDbOk:
		return true
	default:
		return false
	}
}

// Defines values for ReadinessStatusStatus.
const (
	Degraded ReadinessStatusStatus = "degraded"
	Ok       ReadinessStatusStatus = "ok"
)

// Valid indicates whether the value is a known member of the ReadinessStatusStatus enum.
func (e ReadinessStatusStatus) Valid() bool {
	switch e {
	case Degraded:
		return true
	case Ok:
		return true
	default:
		return false
	}
}

// Crawl A crawl job with its full configuration and current status.
type Crawl struct {
	CreatedAt time.Time          `json:"createdAt"`
	Id        openapi_types.UUID `json:"id"`
	MaxDepth  int                `json:"maxDepth"`
	MaxPages  int                `json:"maxPages"`
	SeedUrl   string             `json:"seedUrl"`

	// Status Lifecycle state of a crawl. Only queued is supported for now; more states are added in later specs.
	Status    CrawlStatus `json:"status"`
	UpdatedAt time.Time   `json:"updatedAt"`
}

// CrawlList A page of crawl summaries with an optional cursor to the next page.
type CrawlList struct {
	Items []CrawlSummary `json:"items"`

	// NextCursor Opaque token for the next page, or null when there are no more pages.
	NextCursor *string `json:"nextCursor,omitempty"`
}

// CrawlStatus Lifecycle state of a crawl. Only queued is supported for now; more states are added in later specs.
type CrawlStatus string

// CrawlSummary Condensed crawl representation used in list responses.
type CrawlSummary struct {
	CreatedAt time.Time          `json:"createdAt"`
	Id        openapi_types.UUID `json:"id"`
	SeedUrl   string             `json:"seedUrl"`

	// Status Lifecycle state of a crawl. Only queued is supported for now; more states are added in later specs.
	Status CrawlStatus `json:"status"`
}

// CreateCrawlRequest Parameters for submitting a new crawl job.
type CreateCrawlRequest struct {
	// MaxDepth Maximum link depth to follow from the seed URL.
	MaxDepth int `json:"maxDepth"`

	// MaxPages Maximum number of pages to fetch.
	MaxPages int `json:"maxPages"`

	// SeedUrl Absolute http or https URL to start crawling from.
	SeedUrl string `json:"seedUrl"`
}

// FieldError A single field-level validation failure.
type FieldError struct {
	// Field Name of the request field that failed validation.
	Field string `json:"field"`

	// Message Human-readable explanation of the failure.
	Message string `json:"message"`
}

// HealthStatus Liveness response indicating the process is up and serving requests.
type HealthStatus struct {
	Status HealthStatusStatus `json:"status"`
}

// HealthStatusStatus defines model for HealthStatus.Status.
type HealthStatusStatus string

// NotFoundError Returned with a 404 when a requested resource does not exist.
type NotFoundError struct {
	// Message Human-readable explanation of what was not found.
	Message string `json:"message"`
}

// ReadinessStatus Readiness response reporting aggregate status and per-dependency check results.
type ReadinessStatus struct {
	Checks struct {
		Db ReadinessStatusChecksDb `json:"db"`
	} `json:"checks"`
	Status ReadinessStatusStatus `json:"status"`
}

// ReadinessStatusChecksDb defines model for ReadinessStatus.Checks.Db.
type ReadinessStatusChecksDb string

// ReadinessStatusStatus defines model for ReadinessStatus.Status.
type ReadinessStatusStatus string

// ValidationError Returned with a 400 when request validation fails.
type ValidationError struct {
	Errors []FieldError `json:"errors"`

	// Message Summary of the validation failure.
	Message string `json:"message"`
}

// ListCrawlsParams defines parameters for ListCrawls.
type ListCrawlsParams struct {
	// Cursor Opaque base64url pagination token returned as nextCursor by a previous call.
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Maximum number of items to return.
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// CreateCrawlJSONRequestBody defines body for CreateCrawl for application/json ContentType.
type CreateCrawlJSONRequestBody = CreateCrawlRequest